requirements.txt
attr==0.3.2
beautifulsoup4==4.13.3
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==44.0.0
docutils==0.21.2
HTMLParser==0.0.2
importlib_metadata==8.5.0
importlib_metadata==8.0.0
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.5
jnius==1.1.0
keyring==25.6.0
langchain==0.3.18
langchain_chroma==0.2.1
langchain_community==0.3.17
langchain_core==0.3.34
langchain_groq==0.2.4
langchain_huggingface==0.1.2
lockfile==0.12.2
mock==5.1.0
Pillow==11.1.0
protobuf==5.29.3
pyOpenSSL==25.0.0
python-dotenv==1.0.1
railroad==0.5.0
Sphinx==8.1.3
thread==2.0.5
tornado==6.4.2
truststore==0.10.1
unstructured==0.16.17
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
